Splitted controllers and handlers.

+++ Lapka.Pets.Application/Commands/Handlers/UpdateShelterPetHandler.cs
@@ -11,30 +11,36 @@ namespace Lapka.Pets.Application.Commands.Handlers
     public class UpdatePetHandler : ICommandHandler<UpdateShelterPet>
     {
         private readonly IEventProcessor _eventProcessor;
+        private readonly IPetRepository<ShelterPet> _petRepository;
         private readonly IGrpcPhotoService _grpcPhotoService;
 
-        public UpdatePetHandler(IEventProcessor eventProcessor, IPetRepository<Petd> petRepository, IGrpcPhotoService grpcPhotoService)
+        public UpdatePetHandler(IEventProcessor eventProcessor, IPetRepository<ShelterPet> petRepository,
+            IGrpcPhotoService grpcPhotoService)
         {
             _eventProcessor = eventProcessor;
+            _petRepository = petRepository;
             _grpcPhotoService = grpcPhotoService;
         }
 
         public async Task HandleAsync(UpdateShelterPet command)
         {
-            string mainPhotoPath = $"{Guid.NewGuid():N}.{command.Photo.GetFileExtension()}"; 
+            string mainPhotoPath = $"{Guid.NewGuid():N}.{command.Photo.GetFileExtension()}";
 
-            Pet pet = await _petRepository.GetByIdAsync(command.Id);
+            ShelterPet pet = await _petRepository.GetByIdAsync(command.Id);
             if (pet is null)
             {
                 throw new PetNotFoundException(command.Id);
             }
+
+            pet.Update(command.Name, command.Race, command.Species, mainPhotoPath, command.Sex, command.DateOfBirth,
+                command.Sterilization, command.Weight, command.Color, command.ShelterAddress, command.Description);
+
+            if (command.Photo != null)
+            {
+                await _grpcPhotoService.DeleteAsync(pet.MainPhotoPath);
+                await _grpcPhotoService.AddAsync(mainPhotoPath, command.Photo.Content);
+            }
             
-            pet.Update(command.Name, command.Race, command.Species, mainPhotoPath, command.Sex, command.DateOfBirth, command.Description,
-                command.ShelterAddress, command.Sterilization, command.Weight, command.Color);
-
-            await _grpcPhotoService.DeleteAsync(pet.MainPhotoPath);
-            await _grpcPhotoService.AddAsync(mainPhotoPath, command.Photo.Content);
-
             await _petRepository.UpdateAsync(pet);
             await _eventProcessor.ProcessAsync(pet.Events);
         }
